+++ daemon/images/image.go
@@ -16,7 +16,6 @@ import (
 	"github.com/docker/docker/api/types/backend"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/image"
-	"github.com/docker/docker/layer"
 	"github.com/opencontainers/go-digest"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/pkg/errors"
@@ -163,49 +162,12 @@ func (i *ImageService) manifestMatchesPlatform(ctx context.Context, img *image.I
 	return false, nil
 }
 
-// GetImage returns an image corresponding to the image referred to by refOrID.
-func (i *ImageService) GetImage(ctx context.Context, refOrID string, options backend.GetImageOpts) (*image.Image, error) {
-	img, err := i.getImage(ctx, refOrID, options)
-	if err != nil {
-		return nil, err
-	}
-	if options.Details {
-		var size int64
-		var layerMetadata map[string]string
-		layerID := img.RootFS.ChainID()
-		if layerID != "" {
-			l, err := i.layerStore.Get(layerID)
-			if err != nil {
-				return nil, err
-			}
-			defer layer.ReleaseAndLog(i.layerStore, l)
-			size = l.Size()
-			layerMetadata, err = l.Metadata()
-			if err != nil {
-				return nil, err
-			}
-		}
-
-		lastUpdated, err := i.imageStore.GetLastUpdated(img.ID())
-		if err != nil {
-			return nil, err
-		}
-		img.Details = &image.Details{
-			References:  i.referenceStore.References(img.ID().Digest()),
-			Size:        size,
-			Metadata:    layerMetadata,
-			Driver:      i.layerStore.DriverName(),
-			LastUpdated: lastUpdated,
-		}
-	}
-	return img, nil
-}
-
 func (i *ImageService) GetImageManifest(ctx context.Context, refOrID string, options backend.GetImageOpts) (*ocispec.Descriptor, error) {
 	panic("not implemented")
 }
 
-func (i *ImageService) getImage(ctx context.Context, refOrID string, options backend.GetImageOpts) (retImg *image.Image, retErr error) {
+// GetImage returns an image corresponding to the image referred to by refOrID.
+func (i *ImageService) GetImage(ctx context.Context, refOrID string, options backend.GetImageOpts) (retImg *image.Image, retErr error) {
 	defer func() {
 		if retErr != nil || retImg == nil || options.Platform == nil {
 			return
